Change CCL input precision from int8 to int64 to align with compiler (#1003)
This PR is for changing the precision of the CCL input from int8 to
int64 to be aligned with the compiler for release 1.22. This change is
for addressing the JIRA tickets raised when working with CCL enabled in
new sdk versions.

+++ QEfficient/transformers/models/llava/modeling_llava.py
@@ -213,7 +213,7 @@ class QEffLlavaForConditionalGeneration(LlavaForConditionalGeneration):
         lang_inputs["position_ids"] = torch.full(lang_inputs["position_ids"].shape, CTX_LEN - 1)
 
         if comp_ctx_lengths is not None:
-            lang_inputs["comp_ctx_lengths"] = torch.randint(0, 100, (40,), dtype=torch.int8)
+            lang_inputs["comp_ctx_lengths"] = torch.randint(0, 100, (40,), dtype=torch.int64)
 
         if continuous_batching:
             lang_inputs["batch_index"] = torch.arange(BS).view(BS, 1)
